refactor: decouple CalendarScreen components
